+++ src/Controllers/TriggerController.cs
@@ -1,6 +1,8 @@
-using InvvardDev.Ifttt.Configuration;
 using InvvardDev.Ifttt.Contracts;
+using InvvardDev.Ifttt.Hosting.Models;
 using InvvardDev.Ifttt.Models.Trigger;
+using InvvardDev.Ifttt.Toolkit.Contracts;
+using InvvardDev.Ifttt.Toolkit.Models;
 using Microsoft.AspNetCore.Mvc;
 
 namespace InvvardDev.Ifttt.Controllers;
