+++ src/Reflection/TriggerAttributeLookup.cs
@@ -1,5 +1,6 @@
 using InvvardDev.Ifttt.Contracts;
-using InvvardDev.Ifttt.Models.Attributes;
+using InvvardDev.Ifttt.Toolkit.Attributes;
+using InvvardDev.Ifttt.Toolkit.Contracts;
 
 namespace InvvardDev.Ifttt.Reflection;
