+++ src/Toolkit/Attributes/TriggerAttribute.cs
@@ -1,4 +1,4 @@
-﻿namespace InvvardDev.Ifttt.Models.Attributes;
+﻿namespace InvvardDev.Ifttt.Toolkit.Attributes;
 
 [AttributeUsage(AttributeTargets.Class)]
 public class TriggerAttribute(string slug) : TriggerAttributeBase(slug);
