+++ src/Toolkit/Attributes/TriggerFieldAttribute.cs
@@ -1,4 +1,4 @@
-namespace InvvardDev.Ifttt.Models.Attributes;
+namespace InvvardDev.Ifttt.Toolkit.Attributes;
 
 [AttributeUsage(AttributeTargets.Property)]
 public class TriggerFieldAttribute(string slug) : TriggerAttributeBase(slug);
