+++ src/Toolkit/Contracts/ITrigger.cs
@@ -1,6 +1,6 @@
-﻿using InvvardDev.Ifttt.Models.Trigger;
+﻿using InvvardDev.Ifttt.Toolkit.Models;
 
-namespace InvvardDev.Ifttt.Contracts;
+namespace InvvardDev.Ifttt.Toolkit.Contracts;
 
 public interface ITrigger
 {
