+++ tests/InvvardDev.Ifttt.TestFactories/Triggers/TriggerClassFactory.cs
@@ -1,7 +1,8 @@
 ﻿using InvvardDev.Ifttt.Contracts;
-using InvvardDev.Ifttt.Models.Attributes;
 using InvvardDev.Ifttt.TestFactories.Extensions;
 using InvvardDev.Ifttt.TestFactories.Shared;
+using InvvardDev.Ifttt.Toolkit.Attributes;
+using InvvardDev.Ifttt.Toolkit.Contracts;
 
 namespace InvvardDev.Ifttt.TestFactories.Triggers;
