+++ tests/InvvardDev.Ifttt.Tests/Reflection/TriggerFieldsMapperTests.cs
@@ -1,6 +1,7 @@
 ﻿using FluentAssertions;
-using InvvardDev.Ifttt.Models.Attributes;
-using InvvardDev.Ifttt.Models.Trigger;
+using InvvardDev.Ifttt.Toolkit;
+using InvvardDev.Ifttt.Toolkit.Attributes;
+using InvvardDev.Ifttt.Toolkit.Models;
 
 namespace InvvardDev.Ifttt.Tests.Reflection;
